Implement Load/Save track
There are 2 new folders on /home/pi/.config/LMN-3:
  -  Saved: where all tracks are stored
  -  load_project: the actual project

When the save button is pressed:
  -  the load_project xml file is moved to saved folder, if there are a file whith the same name, it is overwriten
When new track is added:
  -  Load_project folder is cleaned
  -  A new file is created on Load_project folder and the actual project xml file are moved to saved folder, if there are a file whith the same name, it is overwriten
  -  LMN-3 program restart
When a track is loaded:
  -  Load_project folder is cleaned and the track selected is moved to load_project folder
  -  LMN-3 program restart

Since we do not have an alphabetical keyboard, the track name will have the format `edit_+ddmmyyyyhhMMss`.

The file name can be change later, it will appear in the list.